requirements.txt
# diffusers @ git+https://github.com/huggingface/diffusers.git@0d7c4790235ac00b4524b492bc2a680dcc5cf6b0
controlnet_aux==0.0.8
huggingface_hub==0.20.2
insightface==0.7.3
numpy==1.26.2
opencv_contrib_python==4.9.0.80
opencv_python==4.9.0.80
opencv_python_headless==4.7.0.72
Pillow==10.1.0
torch==2.0.0
torchsde==0.2.6
transformers==4.37.1
